backend/pyproject.toml
[project]
name = "hexcoach-backend"
version = "2.0.0"
description = "HexCoach AI Backend API - 英雄联盟 AI 战术副驾系统后端服务"
readme = "../README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "HexCoach Team" }
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.10.0",  # 添加 email 支持
    "python-multipart>=0.0.18",

    # 数据库
    "pymongo>=4.10.0",
    "motor>=3.6.0",           # 异步 MongoDB 驱动
    "beanie>=1.27.0",         # MongoDB ORM

    "python-dotenv>=1.0.0",

    "requests>=2.32.0",

    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    "openai>=1.59.0",

    "edge-tts>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = []
bypass-selection = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "S", "B", "C4", "PIE", "PT"]

ignore = [
    "ANN101",
    "ANN102",
    "S101",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "pymongo.*",
    "edge_tts.*",
    "passlib.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--tb=short",
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    "__pycache__/*",
]

[tool.coverage.report]
precision = 2
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "loguru>=0.7.0",
]
